Restructure Chart in several components
